move dump_loader to dump::loader
